html colors for selected segments in Python API
This feature adds a property: "segment_html_color_dict" to segmentation
layers via the Python API. The property returns a dictionary where
keys are segment ids and values are the html color codes,
e.g. #F9F607 (yellow) corresponding to each currently selected segment.
The colors are the same as what is rendered in the viewer for the segments,
given a color seed. This property uses the current color seed when
calculating the color codes.